Move ellipsoid and geodetic classes into the core
