+++ modules/globebrowsing/src/geojson/globegeometryhelper.cpp
@@ -242,7 +242,7 @@ subdivideTriangle(const glm::vec3& v0, const glm::vec3& v1, const glm::vec3& v2,
     std::vector<Coordinate> pointCoords;
     pointCoords.reserve(3 * maxSteps + 1);
 
-    const globebrowsing::Ellipsoid& ellipsoid = globe.ellipsoid();
+    const Ellipsoid& ellipsoid = globe.ellipsoid();
 
     const float lengthEdge01 = glm::length(v1 - v0);
     const float lengthEdge02 = glm::length(v2 - v0);
